requirements.txt
functions-framework==3.*
opencv-python-headless
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.0.0,<3.0.0